Split EightBittr into separate files
tsconfig.json is being used now, yay.
