+++ tsconfig.json
@@ -1,0 +1,12 @@
+{
+    "compilerOptions": {
+        "module": "es2015",
+        "target": "es2015",
+        "noImplicitAny": true,
+        "sourceMap": true,
+        "outDir": "Distribution"
+    },
+    "exclude": [
+        "node_modules"
+    ]
+}
